Remove ApiExample test base, renamed ApiIntegrationTestBase => ApiTestBase
