+++ src/Tests/Framework/ApiTestBase.cs
@@ -9,13 +9,17 @@ using Elasticsearch.Net;
 
 namespace Tests.Framework
 {
-	public abstract class ApiCallExampleBase<TResponse, TInterface, TDescriptor, TInitializer> : SerializationTestBase
+	public abstract class ApiTestBase<TResponse, TInterface, TDescriptor, TInitializer> 
+		: SerializationTestBase
 		where TResponse : class, IResponse
 		where TDescriptor : class, TInterface
 		where TInitializer : class, TInterface
 		where TInterface : class
 	{
 		private LazyResponses _responses;
+
+		public abstract int ExpectStatusCode { get; }
+		public abstract bool ExpectIsValid { get; }
 
 		public abstract string UrlPath { get; }
 		public abstract HttpMethod HttpMethod { get; }
@@ -26,14 +30,15 @@ namespace Tests.Framework
 
 		readonly IIntegrationCluster _cluster;
 
-		protected ApiCallExampleBase(IIntegrationCluster cluster, EndpointUsage usage)
+		protected abstract LazyResponses ClientUsage();
+
+		protected ApiTestBase(IIntegrationCluster cluster, EndpointUsage usage)
 		{
 			this._cluster = cluster;
 			this.IntegrationPort = cluster.Node.Port;
 			this._responses = usage.CallOnce(this.ClientUsage);
 		}
 
-		protected abstract LazyResponses ClientUsage();
 
 		protected LazyResponses Calls(
 			Func<IElasticClient, Func<TDescriptor, TInterface>, TResponse> fluent,
@@ -57,7 +62,45 @@ namespace Tests.Framework
 				return dict;
 			});
 		}
+		private void AssertUrl(Uri u)
+		{
+			var paths = (this.UrlPath ?? "").Split(new [] { '?' }, 2);
+			string path = paths.First(), query = string.Empty;
+			if (paths.Length > 1)
+				query = paths.Last();
 
+			var expectedUri = new UriBuilder("http","localhost", IntegrationPort, path,  "?" + query).Uri;
+
+			u.AbsolutePath.Should().Be(expectedUri.AbsolutePath);
+			u = new UriBuilder(u.Scheme, u.Host, u.Port, u.AbsolutePath, u.Query.Replace("pretty=true", "")).Uri;
+
+			var queries = new[] {u.Query, expectedUri.Query};
+			if (queries.All(string.IsNullOrWhiteSpace)) return;
+			if (queries.Any(string.IsNullOrWhiteSpace))
+			{
+				queries.Last().Should().Be(queries.First());
+				return;
+			}
+
+			var clientKeyValues = u.Query.Substring(1).Split('&')
+				.Select(v => v.Split('='))
+				.Where(k=>!string.IsNullOrWhiteSpace(k[0]))
+				.ToDictionary(k => k[0], v => v.Last());
+			var expectedKeyValues = expectedUri.Query.Substring(1).Split('&')
+				.Select(v => v.Split('='))
+				.Where(k=>!string.IsNullOrWhiteSpace(k[0]))
+				.ToDictionary(k => k[0], v => v.Last());
+
+			clientKeyValues.Count().Should().Be(expectedKeyValues.Count());
+			clientKeyValues.Should().ContainKeys(expectedKeyValues.Keys.ToArray());
+			clientKeyValues.Should().Equal(expectedKeyValues);
+		}
+
+		[I] protected async Task HandlesStatusCode() =>
+			await this.AssertOnAllResponses(r=>r.ApiCall.HttpStatusCode.Should().Be(this.ExpectStatusCode));
+
+		[I] protected async Task ReturnsExpectedIsValid() =>
+			await this.AssertOnAllResponses(r=>r.IsValid.Should().Be(this.ExpectIsValid));
 		protected static string RandomString() => Guid.NewGuid().ToString("N").Substring(0, 8);
 		protected int IntegrationPort { get; set; } = 9200;
 		protected virtual ConnectionSettings GetConnectionSettings(ConnectionSettings settings) => settings;
@@ -82,40 +125,6 @@ namespace Tests.Framework
 			}
 		}
 
-		private void AssertUrl(Uri u)
-		{
-			var paths = (this.UrlPath ?? "").Split(new[] { '?' }, 2);
-			string path = paths.First(), query = string.Empty;
-			if (paths.Length > 1)
-				query = paths.Last();
-
-			var expectedUri = new UriBuilder("http", "localhost", IntegrationPort, path, "?" + query).Uri;
-
-			u.AbsolutePath.Should().Be(expectedUri.AbsolutePath);
-			u = new UriBuilder(u.Scheme, u.Host, u.Port, u.AbsolutePath, u.Query.Replace("pretty=true", "")).Uri;
-
-			var queries = new[] { u.Query, expectedUri.Query };
-			if (queries.All(string.IsNullOrWhiteSpace)) return;
-			if (queries.Any(string.IsNullOrWhiteSpace))
-			{
-				queries.Last().Should().Be(queries.First());
-				return;
-			}
-
-			var clientKeyValues = u.Query.Substring(1).Split('&')
-				.Select(v => v.Split('='))
-				.Where(k => !string.IsNullOrWhiteSpace(k[0]))
-				.ToDictionary(k => k[0], v => v.Last());
-			var expectedKeyValues = expectedUri.Query.Substring(1).Split('&')
-				.Select(v => v.Split('='))
-				.Where(k => !string.IsNullOrWhiteSpace(k[0]))
-				.ToDictionary(k => k[0], v => v.Last());
-
-			clientKeyValues.Count().Should().Be(expectedKeyValues.Count());
-			clientKeyValues.Should().ContainKeys(expectedKeyValues.Keys.ToArray());
-			clientKeyValues.Should().Equal(expectedKeyValues);
-		}
-
 		[U]
 		protected async Task HitsTheCorrectUrl() =>
 			await this.AssertOnAllResponses(r => this.AssertUrl(r.ApiCall.Uri));
@@ -133,6 +142,6 @@ namespace Tests.Framework
 			this.AssertSerializesAndRoundTrips(this.Fluent(this.ClientDoesThisInternally(NewDescriptor())));
 
 		protected virtual TDescriptor NewDescriptor() => Activator.CreateInstance<TDescriptor>();
-
 	}
+
 }
