+++ src/Tests/Tests.csproj
@@ -99,12 +99,11 @@
     <Compile Include="ClientConcepts\HighLevel\Inferrence\Indices\IndicesPaths.doc.cs" />
     <Compile Include="CodeStandards\Descriptors.doc.cs" />
     <Compile Include="Document\Single\Index\IndexUrlTest.cs" />
-    <Compile Include="Document\Single\Index\Indexing.Doc.cs" />
+    <Compile Include="Document\Single\Index\IndexApiTest.cs" />
     <Compile Include="Framework\CrudTestBase.cs" />
     <Compile Include="Framework\MockResponses\SniffingResponse.cs" />
     <Compile Include="Framework\UrlTests.cs" />
     <Compile Include="Framework\SimpleIntegration.cs" />
-    <Compile Include="Framework\ApiCallExampleBase.cs" />
     <Compile Include="Framework\Audit\Audits.cs" />
     <Compile Include="Framework\Audit\Auditor.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\Sniffing\OnStartup.doc.cs" />
@@ -147,10 +146,10 @@
     <Compile Include="IndexModules\IndexSettings\SlowLog\SlowLogSettings.cs" />
     <Compile Include="IndexModules\IndexSettings\Translog\TranslogSettings.cs" />
     <Compile Include="IndexModules\Similarity\SimilaritySettings.cs" />
-    <Compile Include="Indices\IndexManagement\CreateIndex\CreateIndexFullExample.cs" />
-    <Compile Include="Indices\IndexManagement\CreateIndex\CreateIndex.cs" />
-    <Compile Include="Indices\IndexManagement\PutMapping\PutMapping.cs" />
-    <Compile Include="Indices\IndexManagement\PutMapping\PutMappingFullExample.cs" />
+    <Compile Include="Indices\IndexManagement\CreateIndex\CreateIndexFullApiTest.cs" />
+    <Compile Include="Indices\IndexManagement\CreateIndex\CreateIndexApiTest.cs" />
+    <Compile Include="Indices\IndexManagement\PutMapping\PutMappingApiTest.cs" />
+    <Compile Include="Indices\IndexManagement\PutMapping\PutMappingFullApiTest.cs" />
     <Compile Include="Mapping\Types\String.cs" />
     <Compile Include="Search\Request\Fields.cs" />
     <Compile Include="Search\Request\FielddataFields.cs" />
@@ -166,7 +165,7 @@
     <Compile Include="Search\Request\FromAndSize.cs" />
     <Compile Include="Search\SearchUsageBase.cs" />
     <Compile Include="Framework\AsyncLazy.cs" />
-    <Compile Include="Framework\ApiCallIntegrationBase.cs" />
+    <Compile Include="Framework\ApiTestBase.cs" />
     <Compile Include="Framework\UsageTestBase.cs" />
     <Compile Include="Framework\Integration\Seeder.cs" />
     <Compile Include="Framework\MockData\Project.cs" />
